requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
supabase>=2.0.0
openpyxl>=3.1.0
toml>=0.10.2
argon2-cffi>=23.1.0
sib-api-v3-sdk>=7.6.0
openai>=1.3.0
xmltodict>=0.13.0
extra-streamlit-components>=0.1.60
PyMuPDF>=1.23.0
Pillow>=10.0.0
requests>=2.31.0
tenacity>=8.2.3
asn1crypto>=1.5.1
# Security: fixes CVE-2026-26007
cryptography>=46.0.5
# Security: fixes CVE-2026-32597
PyJWT>=2.12.0
# FastAPI worker (Fase 3 — AI/parsing separato da Streamlit)
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.12
python-dotenv>=1.0.0
